+++ core/src/main/java/org/web3j/protocol/core/DefaultIdProvider.java
@@ -17,8 +17,7 @@ import java.util.concurrent.atomic.AtomicLong;
 public class DefaultIdProvider {
     protected static final AtomicLong nextId = new AtomicLong(0);
 
-    protected DefaultIdProvider() {
-    }
+    protected DefaultIdProvider() {}
 
     public static long getNextId() {
         return nextId.getAndIncrement();