+++ core/src/test/java/org/web3j/protocol/core/RequestTest.java
@@ -843,4 +843,48 @@ class RequestTest extends RequestTester {
         BigInteger resultWhenExcessBlobGasIsNotZero = jsonRpc20Web3j.ethGetBaseFeePerBlobGas();
         assertEquals(BigInteger.valueOf(21518435987L), resultWhenExcessBlobGasIsNotZero);
     }
+
+    @Test
+    void testLineaEstimateGas() throws Exception {
+        web3j.lineaEstimateGas(
+                        Transaction.createEthCallTransaction(
+                                "0xa70e8dd61c5d32be8058bb8eb970870f07233155",
+                                "0x52b93c80364dc2dd4444c146d73b9836bbbb2b3f",
+                                "0x0"))
+                .send();
+
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"linea_estimateGas\","
+                        + "\"params\":[{\"from\":\"0xa70e8dd61c5d32be8058bb8eb970870f07233155\","
+                        + "\"to\":\"0x52b93c80364dc2dd4444c146d73b9836bbbb2b3f\",\"data\":\"0x0\"}],"
+                        + "\"id\":1}");
+    }
+
+    @Test
+    void testLineaGetProof() throws Exception {
+        web3j.lineaGetProof(
+                        "0x7F0d15C7FAae65896648C8273B6d7E43f58Fa842",
+                        Arrays.asList(
+                                "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421"),
+                        "latest")
+                .send();
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"linea_getProof\","
+                        + "\"params\":[\"0x7F0d15C7FAae65896648C8273B6d7E43f58Fa842\","
+                        + "[\"0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421\"],"
+                        + "\"latest\"],"
+                        + "\"id\":0}");
+    }
+
+    @Test
+    void testLineaGetTransactionExclusionStatusV1() throws Exception {
+        web3j.lineaGetTransactionExclusionStatusV1(
+                        "0xb903239f8543d04b5dc1ba6579132b143087c68db1b2168786408fcbce568238")
+                .send();
+
+        verifyResult(
+                "{\"jsonrpc\":\"2.0\",\"method\":\"linea_getTransactionExclusionStatusV1\",\"params\":["
+                        + "\"0xb903239f8543d04b5dc1ba6579132b143087c68db1b2168786408fcbce568238\"],"
+                        + "\"id\":1}");
+    }
 }