+++ CHANGELOG.md
@@ -3,7 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [4.12.4]() (Upcoming)
+# [4.13.1]() (2025-03-05)
+
+### Bug Fixes
+
+* 
+### Features
+
+* bump snapshot version to 4.13.1 [#2160](https://github.com/hyperledger-web3j/web3j/pull/2160)
+
+
+### BREAKING CHANGES
+
+*
+
+# [4.13.0](https://github.com/hyperledger-web3j/web3j/releases/tag/v4.13.0) (2025-03-05)
 
 ### Bug Fixes